+++ plugins/alignments/src/LinearReadArcsDisplay/model.tsx
@@ -18,7 +18,11 @@ import { FilterModel } from '../shared'
 import drawFeats from './drawFeats'
 import { fetchChains, ChainData } from '../shared/fetchChains'
 import { BaseDisplay } from '@jbrowse/core/pluggableElementTypes'
-import { HeightMixin } from '@jbrowse/plugin-linear-genome-view'
+import {
+  FeatureDensityMixin,
+  TrackHeightMixin,
+} from '@jbrowse/plugin-linear-genome-view'
+import { createAutorun } from '../util'
 
 // async
 const FilterByTagDlg = lazy(() => import('../shared/FilterByTag'))
@@ -39,7 +43,8 @@ function stateModelFactory(configSchema: AnyConfigurationSchemaType) {
     .compose(
       'LinearReadArcsDisplay',
       BaseDisplay,
-      HeightMixin(),
+      TrackHeightMixin(),
+      FeatureDensityMixin(),
       types.model({
         /**
          * #property
@@ -91,22 +96,20 @@ function stateModelFactory(configSchema: AnyConfigurationSchemaType) {
       loading: false,
       drawn: false,
       chainData: undefined as ChainData | undefined,
-      message: '',
       ref: null as HTMLCanvasElement | null,
       lastDrawnOffsetPx: 0,
     }))
-
     .actions(self => ({
       /**
        * #action
-       * internal, a reference to a HTMLCanvas because we use a autorun to draw the canvas
        */
       reload() {
         self.error = undefined
       },
       /**
        * #action
-       * internal, a reference to a HTMLCanvas because we use a autorun to draw the canvas
+       * internal, a reference to a HTMLCanvas because we use a autorun to draw
+       * the canvas
        */
       setRef(ref: HTMLCanvasElement | null) {
         self.ref = ref
@@ -277,7 +280,6 @@ function stateModelFactory(configSchema: AnyConfigurationSchemaType) {
                 },
               ],
             },
-
             {
               label: 'Draw inter-region vertical lines',
               type: 'checkbox',
@@ -335,31 +337,24 @@ function stateModelFactory(configSchema: AnyConfigurationSchemaType) {
           autorun(() => fetchChains(self), { delay: 1000 }),
         )
 
-        addDisposer(
+        createAutorun(
           self,
-          autorun(
-            async () => {
-              try {
-                const canvas = self.ref
-                if (!canvas) {
-                  return
-                }
-                const ctx = canvas.getContext('2d')
-                if (!ctx) {
-                  return
-                }
-                ctx.clearRect(0, 0, canvas.width, self.height * 2)
-                ctx.resetTransform()
-                ctx.scale(2, 2)
-                await drawFeats(self, ctx)
-                self.setDrawn(true)
-              } catch (e) {
-                console.error(e)
-                self.setError(e)
-              }
-            },
-            { delay: 1000 },
-          ),
+          async () => {
+            const canvas = self.ref
+            if (!canvas) {
+              return
+            }
+            const ctx = canvas.getContext('2d')
+            if (!ctx) {
+              return
+            }
+            ctx.clearRect(0, 0, canvas.width, self.height * 2)
+            ctx.resetTransform()
+            ctx.scale(2, 2)
+            drawFeats(self, ctx)
+            self.setDrawn(true)
+          },
+          { delay: 1000 },
         )
       },
     }))
