+++ plugins/linear-genome-view/src/index.ts
@@ -66,7 +66,8 @@ export {
   BaseLinearDisplay,
   BlockMsg,
   BaseLinearDisplayComponent,
-  HeightMixin,
+  TrackHeightMixin,
+  FeatureDensityMixin,
 } from './BaseLinearDisplay'
 export {
   type LinearGenomeViewModel,
